Further refactoring, export/import of configs is being implemented. With the refactoring, #25 should be done. - closes #25 #27 is being worked on. - closes #30 - closes #37 #45 is in process
