+++ pyproject.toml
@@ -7,12 +7,13 @@ authors = ["None"]
 [tool.poetry.dependencies]
 python = "^3.9"
 linkfinder = { path = "./src/linkfinder" }
-chardet = "5.0.0"
-pycurl = "7.45.1"
-pyparsing = "2.3.0"
-pySocks = "1.7.1"
+chardet = "^5.0.0"
+pycurl = "^7.45.1"
+pyparsing = "^2.3.0"
+pySocks = "^1.7.1"
 six = "^1.11.0"
 requests = "^2.28.1"
+tomlkit = "^0.12.1"
 beautifulsoup4 = "^4.11.1"
 python-dateutil = "^2.8.2"
 pytz = "^2022.6"
