+++ src/wenum/plugins/scripts/listing.py
@@ -3,7 +3,7 @@ import re
 from wenum.plugin_api.base import BasePlugin
 from wenum.plugin_api.static_data import LISTING_dir_indexing_regexes
 from wenum.externals.moduleman.plugin import moduleman_plugin
-from wenum.ui.console.common import Term
+from wenum.ui.console.term import Term
 
 
 @moduleman_plugin
