+++ src/wenum/plugins/scripts/title.py
@@ -1,6 +1,6 @@
 from wenum.plugin_api.base import BasePlugin
 from wenum.externals.moduleman.plugin import moduleman_plugin
-from wenum.ui.console.common import Term
+from wenum.ui.console.term import Term
 
 
 @moduleman_plugin
