+++ src/wenum/ui/console/kbhit.py
@@ -36,7 +36,6 @@ class KBHit:
             # Support normal-terminal reset at exit
             atexit.register(self.set_normal_term)
 
-
     def set_normal_term(self):
         """
         Resets to normal terminal. On Windows this is a no-op.
@@ -48,13 +47,10 @@ class KBHit:
         else:
             termios.tcsetattr(self.fd, termios.TCSAFLUSH, self.old_term)
 
-
     def getch(self):
-        ''' Returns a keyboard character after kbhit() has been called.
+        """ Returns a keyboard character after kbhit() has been called.
             Should not be called in the same program as getarrow().
-        '''
-
-        s = ''
+        """
 
         if os.name == 'nt':
             return msvcrt.getch().decode('utf-8')
@@ -62,15 +58,14 @@ class KBHit:
         else:
             return sys.stdin.read(1)
 
-
     def getarrow(self):
-        ''' Returns an arrow-key code after kbhit() has been called. Codes are
+        """ Returns an arrow-key code after kbhit() has been called. Codes are
         0 : up
         1 : right
         2 : down
         3 : left
         Should not be called in the same program as getch().
-        '''
+        """
 
         if os.name == 'nt':
             msvcrt.getch() # skip 0xE0
@@ -83,10 +78,9 @@ class KBHit:
 
         return vals.index(ord(c.decode('utf-8')))
 
-
     def kbhit(self):
-        ''' Returns True if keyboard character was hit, False otherwise.
-        '''
+        """ Returns True if keyboard character was hit, False otherwise.
+        """
         if os.name == 'nt':
             return msvcrt.kbhit()
